Added in basic movement + camera
Currently added in base movement, a camera, and a dummy avatar with an idle and walk animation set up.

I added in a simple state machine, which, while not completely useful right now, could be beneficial when we add in more actions such as holding the lighter. It will also come in handy when implementing enemies.

Current models provided by marianpekar:
https://github.com/marianpekar/unity-third-person-controller