Animation modifications with lighter
Made some changes to animations to allow for there to be a lighter holding animation. I am currently substituted a default Run_Shoot animation as the lighter holding animation, but will need to be changed for the future character model. I used a skeleton mask to allow animation merging, so that the player can run and also hold something without the two animations colliding.

I also adjusted the lighter light to follow the hand so the new model will be easier to add in.