Created object pickup
Created a simple script to allow you to "pick up" objects using the designated interact button/control.
The "pick up" does not do much right now as it simply removes the object from the game. I did not feel the need to actually store it because we aren't worrying about inventory.