Bug fixes + rig setup
Got the rig working with the bones and basic animations, although they are still buggy right now. There were problems with the scripts that relied on the Player being named Player, which was changed in the prefab, so they had to be slightly reworked.

The animations for Harry have been only barely implemented, as the current ones are buggy and don't loop properly.

Other than that, there were just other small bug changes.